Resolve tests, optional imports
