+++ .github/workflows/python-test.yaml
@@ -20,7 +20,7 @@ jobs:
           python -m pip install --upgrade pip
           pip install ruff pytest
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-          pip install datasets torch nltk openai
+          pip install selenium nltk
       - name: Test with pytest
         run: |
           pytest
