+++ examples_py/persona_chat_example/components/metrics.py
@@ -3,8 +3,10 @@ from __future__ import annotations
 import re
 from collections import Counter
 from dataclasses import dataclass
+from typing import TYPE_CHECKING
 
-import torch
+if TYPE_CHECKING:
+    import torch
 
 
 class Perplexity:
