+++ requirements.txt
@@ -1,4 +1,3 @@
 Flask
 Flask-Cors
-selenium
 requests
